+++ src/database/db.ts
@@ -1,3 +1,0 @@
-import Sqlite from "better-sqlite3";
-const sql = Sqlite("./database/blackhole.sqlite", { fileMustExist: false });
-export default sql;