+++ CHANGELOG.md
@@ -27,5 +27,11 @@
 
 * Updated the file & folder structure for the entire project
 * Add a nice way to build the project inside the `/build` directory
-* Built version of blackhole now runs in less than 50MB after a minute or so
+* Compiled version of blackhole now runs in less than 50MB after a minute or so
 ![blackhole_performance_under 50MB_memory](./tmp/blackhole_performance.png)
+
+## v0.1.0
+
+* Removed all custom css and replaced with bootstrap
+* Made block/allowed list to be compactable with import from different domain lists
+